Move factories to `API\Repositories\Factories`
